renamed comments package in display post from list to comments
