+++ app/src/main/java/com/relic/presentation/displaypost/comments/CommentMoreItemsVH.kt
@@ -1,4 +1,4 @@
-package com.relic.presentation.displaypost.list
+package com.relic.presentation.displaypost.comments
 
 import android.support.v7.widget.RecyclerView
 import com.relic.domain.models.CommentModel
