+++ app/src/main/java/com/relic/presentation/displaypost/comments/FullPostVH.kt
@@ -1,10 +1,9 @@
-package com.relic.presentation.displaypost.list
+package com.relic.presentation.displaypost.comments
 
 import android.support.v7.widget.RecyclerView
 import com.relic.domain.models.PostModel
 import com.relic.presentation.displaypost.DisplayPostContract
 import com.relic.presentation.displaypost.FullPostView
-import com.relic.presentation.util.MediaHelper
 
 class FullPostVH(
     private val fullPost : FullPostView
