+++ app/src/main/java/com/relic/presentation/displaypost/tabs/CommentsFragment.kt
@@ -9,9 +9,8 @@ import android.view.View
 import android.view.ViewGroup
 import com.relic.R
 import com.relic.presentation.base.RelicFragment
-import com.relic.presentation.displaypost.DisplayPostFragment
 import com.relic.presentation.displaypost.DisplayPostVM
-import com.relic.presentation.displaypost.list.CommentItemAdapter
+import com.relic.presentation.displaypost.comments.CommentItemAdapter
 import com.shopify.livedataktx.nonNull
 import com.shopify.livedataktx.observe
 import kotlinx.android.synthetic.main.tab_comments.*
@@ -49,10 +48,7 @@ class CommentsFragment : RelicFragment() {
         super.bindViewModel(lifecycleOwner)
 
         commentsVM.commentListLiveData.nonNull().observe(lifecycleOwner) {
-            commentAdapter.setComments(it) {
-                (parentFragment as DisplayPostFragment).onPostDataLoaded()
-            }
-
+            commentAdapter.setComments(it)
             commentsTabSwipeRefresh.isRefreshing = false
         }
     }
